rename flint2 -> flint
